renamed middleware dir to handlers
