+++ src/middleware/services/upload.service.ts
@@ -3,9 +3,9 @@ import { Service } from '../../core/app';
 import { RouteProperties, Transform } from '../../core/interface';
 import { method } from '../guards/method';
 import { pathGuard } from '../guards/path';
-import { saveFiles, SaveFilesProperties } from '../middleware/saveFiles';
-import { serve } from '../middleware/serve';
-import { serveFile } from '../middleware/serveFile';
+import { saveFiles, SaveFilesProperties } from '../handlers/saveFiles';
+import { serve } from '../handlers/serve';
+import { serveFile } from '../handlers/serveFile';
 import { fileProcessor } from '../processors/before/file.processor';
 import { directoryTransform } from '../transforms/directory.transform';
 import { htmlTemplate } from '../util/htmlTemplate';
